// src/app/interfaces.ts
export interface Label {
    _id: string;
    color: string;
    createdAt: number;
    groupId: string;
    icon: string;
    isAction: boolean;
    isHidden: boolean;
    showAs: "text" | "icon" | "both";
    title: string;
}

export interface LabelGroup {
    _id: string;
    color?: string;
    createdAt: number;
    icon?: string;
    isExclusive: boolean;
    isMenu?: boolean; // optional, as it is no longer used
    rank: number;
    title: string;
}

export interface Project {
    _id: string;
    color: string;
    day: string | null | undefined;
    dayRank: number;
    db: string;
    done: boolean;
    doneDate: string;
    dueDate: string;
    echo: boolean;
    endDate: string;
    firstScheduled: string | "unassigned";
    icon: string;
    isFrogged: boolean;
    labelIds: string[];
    marvinPoints: number;
    mpNotes: string[];
    note: string;
    parentId: string;
    plannedMonth: string;
    plannedWeek: string;
    priority: "low" | "mid" | "high";
    rank: number;
    recurring: boolean;
    recurringTaskId: string;
    reviewDate: string;
    sprintId: string;
    startDate: string;
    timeEstimate: number;
    title: string;
    type: "project" | "category";
    updatedAt: number;
    workedOnAt: number;
}
// ...
export interface Subtask {
  _id: string;
  done: boolean;
  rank: number;
  timeEstimate: number;
  title: string;

  // NEW REMINDER FORMAT
  autoSnooze?: boolean;
  reminderOffset?: number;
  reminderTime?: number;
  snooze?: number;
  taskTime?: string;

  // OLD REMINDER FORMAT
  remindAt?: string;
  reminder?: {
    time: string;
    diff: number;
  };
}

export interface Task {
  _id: string;
  backburner: boolean;
  bonusSection: string;
  calData: string;
  calId: string;
  calURL: string;
  colorBar: string | null;
  completedAt: number | null;
  createdAt: number;
  customSection: string;
  dailySection: string;
  day: string;
  db: string;
  deletedAt: number;
  dependsOn: { [id: string]: boolean };
  done: boolean;
  doneAt: number;
  dueDate: string | null;
  duration: number;
  echo: boolean;
  echoedAt: number;
  echoId: string;
  email: string;
  endDate: string | null;
  etag: string;
  firstScheduled: string;
  firstTracked: number;
  generatedAt: number;
  imported: boolean;
  isFrogged: boolean;
  isPinned: boolean;
  isReward: boolean;
  isStarred: boolean;
  itemSnoozeTime: number;
  labelIds: string[];
  link: string;
  masterRank: number;
  note: string;
  onboard: boolean;
  parentId: string;
  permaSnoozeTime: string;
  pinId: string;
  plannedMonth: string;
  plannedWeek: string;
  rank: number;
  recurring: boolean;
  recurringTaskId: string;
  restoredAt: number;
  reviewDate: string;
  sprintId: string | null;
  startDate: string | null;
  subtasks: { [id: string]: Subtask };
  timeBlockSection: string;
  timeEstimate: number;
  times: number[];
  title: string;
  updatedAt: number;
  workedOnAt: number;

  // Gamification
  marvinPoints?: number;
  mpNotes?: string[];
  rewardId?: number;
  rewardPoints?: number;

  // Goals
  g_in_GOALID?: boolean;
  g_sec_GOALID?: string;
  g_rank_GOALID?: number;

  // NEW REMINDER FORMAT
  autoSnooze?: boolean;
  reminderOffset?: number;
  reminderTime?: number;
  snooze?: number;
  taskTime?: string;

  // OLD REMINDER FORMAT
  remindAt?: string;
  reminder?: {
    time: string;
    diff: number;
  };
}